+++ integration_test/tests/index.ts
@@ -1,6 +1,13 @@
 import AdvertisingTest from './advertisingTest';
+import CaptionTest from './captionTest';
 import LoadingTest from './loadingTest';
 import PlaybackTest from './playbackTest';
 import UnloadingTest from './unloadingTest';
 
-export default [AdvertisingTest, LoadingTest, PlaybackTest, UnloadingTest];
+export default [
+  AdvertisingTest,
+  CaptionTest,
+  LoadingTest,
+  PlaybackTest,
+  UnloadingTest,
+];